fix: proper validation of asset unlock tx
Asset unlock txes should be fully validated not only once added to mempool,
but every time when they are candidates to be mined in new block.
Otherwise block can contains invalid transaction.

Example, when it can happen. Tx is valid only until quorum is active (or during to next one).
Once tx is added to mempool, height of block is not validated anymore by previous implementation.

If unlock tx is re-validated during block mining it may be discarded if wasn't mined for long time