refactor: convert maybe_error into generic `Result` template class
This is a simple implementation of the Result type in C++. The Result type is a type that can either be an Ok value, which represents a successful operation and contains a value, or an Err value, which represents a failed operation and contains an error value.

The Result class is a template class that takes two types: T, which is the type of the value that an Ok result will contain, and E, which is the type of the value that an Err result will contain.

The Ok and Err structs are also templates, and are used to construct Ok and Err values. The Ok struct has a single member val_ of type T, and the Err struct has a single member val_ of type E.

The Result class has a constructor that takes either an Ok value or an Err value, and stores the value internally. It also has several member functions that can be used to access the value of the Result and check whether it is an Ok or Err value.

The is_ok() and is_err() member functions can be used to check whether the Result is an Ok or Err value, respectively. The operator bool() function allows the Result to be used in a boolean context, such as an if statement. The unwrap() and unwrap_err() member functions can be used to get the value contained in the Result, but will assert if the Result is not of the expected type.